+++ ChangeLog.md
@@ -6,6 +6,8 @@
   * Add string input element (Mariana Silva).
 
   * Add element to display matrix in latex format (Mariana Silva).
+  * Add string input element (Mariana Silva)
+  * Add matrix grid input element (Mariana Silva)
 
   * Add student name and clickable e-mail address information to issue reports (James Balamuta).